+++ StripeCore/StripeCore/Source/API Bindings/StripeAPIConfiguration+Version.swift
@@ -10,7 +10,7 @@ import Foundation
 
 public extension StripeAPIConfiguration {
     /// The current version of this library.
-    static let STPSDKVersion = "22.8.4"
+    static let STPSDKVersion = "23.0.0"
 
     /*
      NOTE: `STPSDKVersion` must be a hard-coded static string instead of